src/main/java/com/phamhieu/bookapi/api/book/BookDTOMapper.java
package com.phamhieu.bookapi.api.book;

import com.phamhieu.bookapi.domain.book.Book;

import java.util.List;

import static org.apache.commons.collections4.ListUtils.emptyIfNull;

public class BookDTOMapper {
    public static BookDTO toBookDTO(final Book book) {
        return BookDTO.builder()
                .id(book.getId())
                .title(book.getTitle())
                .author(book.getAuthor())
                .description(book.getDescription())
                .createdAt(book.getCreatedAt())
                .updatedAt(book.getUpdatedAt())
                .image(book.getImage())
                .userId(book.getUserId())
                .build();
    }

    public static List<BookDTO> toBookDTOs(final List<Book> books) {
        return emptyIfNull(books)
                .stream()
                .map(BookDTOMapper::toBookDTO)
                .toList();
    }

    public static Book toBook(final BookDTO bookDTO) {
        return Book.builder()
                .id(bookDTO.getId())
                .title(bookDTO.getTitle())
                .author(bookDTO.getAuthor())
                .description(bookDTO.getDescription())
                .createdAt(bookDTO.getCreatedAt())
                .updatedAt(bookDTO.getUpdatedAt())
                .image(bookDTO.getImage())
                .userId(bookDTO.getUserId())
                .build();
    }
}
